std: reorganize the UNIX-internal `weak` module
