+++ library/std/src/sys/pal/unix/weak/tests.rs
@@ -1,30 +1,24 @@
-use super::*;
+// This file is included by both implementations of `weak!`.
+use super::weak;
+use crate::ffi::{CStr, c_char};
 
 #[test]
-fn dlsym_existing() {
+fn weak_existing() {
     const TEST_STRING: &'static CStr = c"Ferris!";
 
     // Try to find a symbol that definitely exists.
-    dlsym! {
+    weak! {
         fn strlen(cs: *const c_char) -> usize;
-    }
-
-    dlsym! {
-        #[link_name = "strlen"]
-        fn custom_name(cs: *const c_char) -> usize;
     }
 
     let strlen = strlen.get().unwrap();
     assert_eq!(unsafe { strlen(TEST_STRING.as_ptr()) }, TEST_STRING.count_bytes());
-
-    let custom_name = custom_name.get().unwrap();
-    assert_eq!(unsafe { custom_name(TEST_STRING.as_ptr()) }, TEST_STRING.count_bytes());
 }
 
 #[test]
-fn dlsym_missing() {
+fn weak_missing() {
     // Try to find a symbol that definitely does not exist.
-    dlsym! {
+    weak! {
         fn test_symbol_that_does_not_exist() -> i32;
     }
